backend/requirements.txt
chess
fastapi
uvicorn
sqlalchemy
google-genai
chromadb
berserk